requirements.txt
flask==2.2.1
requests==2.28.0
selenium==4.4.2
scrapetube==2.2.2
google-api-python-client==2.58.0
pandas==1.4.3
mysql-connector-python==8.0.28
selenium==4.4.2
pytube==12.1.0
pymongo==3.9.0
pdfkit==1.0.0
lxml==4.9.1